Delete actualizado para persona y profesion

+++ personapi-dotnet/personapi-dotnet/Controllers/PersonaController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using personapi_dotnet.Models.Entities;
 using personapi_dotnet.Models.Interfaces;
 
@@ -7,10 +8,12 @@ namespace personapi_dotnet.Controllers
 	public class PersonaController : Controller
 	{
 		private readonly IPersonaRepository _personaRepo;
+		private readonly persona_dbContext _context;
 
-		public PersonaController(IPersonaRepository personaRepo)
+		public PersonaController(IPersonaRepository personaRepo, persona_dbContext context)
 		{
 			_personaRepo = personaRepo;
+			_context = context;
 		}
 
 		public async Task<IActionResult> Index()
@@ -81,10 +84,22 @@ namespace personapi_dotnet.Controllers
 		public async Task<IActionResult> DeleteConfirmed(int id)
 		{
 			var persona = await _personaRepo.GetByIdAsync(id);
+
+			if (persona == null)
+				return NotFound();
+
+			var estudios = await _context.Estudios.Where(e => e.CcPer == id).ToListAsync();
+			_context.Estudios.RemoveRange(estudios);
+
+			var telefonos = await _context.Telefonos.Where(t => t.Duenio == id).ToListAsync();
+			_context.Telefonos.RemoveRange(telefonos);
+
 			_personaRepo.Delete(persona);
 			await _personaRepo.SaveAsync();
+
 			return RedirectToAction(nameof(Index));
 		}
+
 
 		public async Task<IActionResult> Details(int id)
 		{
